Get rid of control version
