+++ frontend/src/scenes/billing/Billing.tsx
@@ -4,14 +4,11 @@ import { CurrentUsage } from './CurrentUsage'
 import { BillingEnrollment } from './BillingEnrollment'
 import './Billing.scss'
 import { billingLogic } from './billingLogic'
-import { featureFlagLogic } from 'lib/logic/featureFlagLogic'
-import { FEATURE_FLAGS } from 'lib/constants'
 import { SceneExport } from 'scenes/sceneTypes'
 import { LemonButton, LemonDivider } from '@posthog/lemon-ui'
 import { AlertMessage } from 'lib/components/AlertMessage'
 import { useValues } from 'kea'
-import { BillingV2 } from './v2/control/Billing'
-import { BillingV2 as BillingV2Test } from './v2/test/Billing'
+import { BillingV2 } from './v2/Billing'
 import { SpinnerOverlay } from 'lib/components/Spinner/Spinner'
 
 export const scene: SceneExport = {
@@ -21,7 +18,6 @@ export const scene: SceneExport = {
 
 export function Billing(): JSX.Element {
     const { billing, isSmallScreen, billingVersion } = useValues(billingLogic)
-    const { featureFlags } = useValues(featureFlagLogic)
 
     if (!billingVersion) {
         return <SpinnerOverlay />
@@ -31,7 +27,7 @@ export function Billing(): JSX.Element {
         return (
             <div>
                 <PageHeader title="Billing &amp; usage" />
-                {featureFlags[FEATURE_FLAGS.BILLING_FEATURES_EXPERIMENT] === 'test' ? <BillingV2Test /> : <BillingV2 />}
+                <BillingV2 />
             </div>
         )
     }
