+++ frontend/src/scenes/billing/v2/BillingGauge.scss
