+++ frontend/src/scenes/billing/v2/BillingHero.scss
