+++ frontend/src/scenes/billing/v2/PlanTable.scss
