--- frontend/src/scenes/billing/v2/test/billing-utils.ts
@@ -1,97 +1,0 @@
-import { dayjs } from 'lib/dayjs'
-import { BillingProductV2Type, BillingV2Type } from '~/types'
-
-export const summarizeUsage = (usage: number | null): string => {
-    if (usage === null) {
-        return ''
-    } else if (usage < 1000) {
-        return `${usage}`
-    } else if (Math.round(usage / 1000) < 1000) {
-        return `${Math.round(usage / 1000)} thousand`
-    } else {
-        return `${Math.round(usage / 1000000)} million`
-    }
-}
-
-export const projectUsage = (
-    usage: number | undefined,
-    period: BillingV2Type['billing_period']
-): number | undefined => {
-    if (typeof usage === 'undefined') {
-        return usage
-    }
-    if (!period) {
-        return usage
-    }
-
-    const timeSoFar = dayjs().diff(period.current_period_start, 'hours')
-
-    // If less than 6 hours have passed, we don't have enough data to project
-    if (timeSoFar <= 6) {
-        return usage
-    }
-    const timeTotal = period.current_period_end.diff(period.current_period_start, 'hours')
-
-    return Math.round((usage / timeSoFar) * timeTotal)
-}
-
-export const convertUsageToAmount = (usage: number, tiers: BillingProductV2Type['tiers']): string => {
-    let remainingUsage = usage
-    let amount = 0
-    let previousTier: BillingProductV2Type['tiers'][0] | undefined = undefined
-
-    for (const tier of tiers) {
-        if (remainingUsage <= 0) {
-            break
-        }
-
-        const tierUsageMax = tier.up_to ? tier.up_to - (previousTier?.up_to || 0) : Infinity
-        const amountFloatUsd = parseFloat(tier.unit_amount_usd)
-        const usageThisTier = Math.min(remainingUsage, tierUsageMax)
-        remainingUsage -= usageThisTier
-        amount += amountFloatUsd * usageThisTier
-        previousTier = tier
-    }
-
-    return amount.toFixed(2)
-}
-
-export const convertAmountToUsage = (amount: string, tiers: BillingProductV2Type['tiers']): number => {
-    if (!amount) {
-        return 0
-    }
-
-    let remainingAmount = parseFloat(amount)
-    let usage = 0
-    let previousTier: BillingProductV2Type['tiers'][0] | undefined = undefined
-
-    if (remainingAmount === 0) {
-        if (parseFloat(tiers[0].unit_amount_usd) === 0) {
-            return tiers[0].up_to || 0
-        }
-        return 0
-    }
-
-    const allTiersZero = tiers.every((tier) => !parseFloat(tier.unit_amount_usd))
-
-    if (allTiersZero) {
-        // Free plan - usage cannot be calculated
-        return tiers[0].up_to || 0
-    }
-
-    for (const tier of tiers) {
-        if (remainingAmount <= 0) {
-            break
-        }
-
-        const tierUsageMax = tier.up_to ? tier.up_to - (previousTier?.up_to || 0) : Infinity
-        const amountFloatUsd = parseFloat(tier.unit_amount_usd)
-        const usageThisTier = Math.min(remainingAmount / amountFloatUsd, tierUsageMax)
-
-        usage += usageThisTier
-        remainingAmount -= amountFloatUsd * usageThisTier
-        previousTier = tier
-    }
-
-    return Math.round(usage)
-}
